attached_assets/requirements_rag_1760607549721.txt
# Core Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.2.0

# LLM & RAG Framework
anthropic==0.18.1
openai==1.12.0
langchain==0.1.10
langchain-anthropic==0.1.4
langchain-openai==0.0.6
langchain-community==0.0.24

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Agentic Framework
langgraph==0.0.20

# Evaluation
ragas==0.1.5
datasets==2.16.1

# Data Processing
numpy==1.26.3
pandas==2.2.0

# Environment & Configuration
python-dotenv==1.0.1

# Optional: External Search API
# tavily-python==0.3.0

# Development & Testing
pytest==8.0.0
pytest-asyncio==0.23.4
